src/main/java/com/example/testtask/controller/ArticleController.java
package com.example.testtask.controller;

import com.example.testtask.model.Article;
import com.example.testtask.service.ArticleService;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api")
public class ArticleController {
    private final ArticleService articleService;

    public ArticleController(ArticleService articleService) {
        this.articleService = articleService;
    }

    @PostMapping("/articles/save")
    public Article saveArticle(@RequestBody Article article) {
        return articleService.save(article);
    }
}
